Restructure context file structure to prepare for adding stack context
